fix: move scripts from _document to _app
